Fix bugs on page permissions
