+++ src/permissions/can.js
@@ -1,9 +1,24 @@
 import { useState, useMemo } from 'react'
 import P from 'prop-types'
 
-import check from './check'
+import rules from './rules'
 
 const systemType = process.env.NEXT_PUBLIC_SYSTEM_TYPE
+
+const check = (type, role, action) => {
+  const permissions = rules[type][role].actions
+
+  if (!permissions) {
+    console.log("ERROR: one of the provided params don't exist!")
+    return false
+  }
+
+  if (permissions.includes(action)) {
+    return true
+  }
+
+  return false
+}
 
 const Can = ({ perform, yes, no }) => {
   const [user] = useState(
